+++ app/Http/Controllers/TodosController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\TodoRequest;
 use Illuminate\Http\Request;
 use App\Models\Todo;
 use App\Repositories\Interfaces\TodoRepositoryInterface;
+use Illuminate\Support\Facades\Auth;
 
 class TodosController extends Controller
 {
@@ -48,11 +49,16 @@ class TodosController extends Controller
     {
         $data = $request->all();
         $todo = new Todo();
+        $todo->user_id = Auth::user()->id;
         $todo->name = $data['name'];
         $todo->description = $data['description'];
         $todo->completed = false;
-        $todo->save();
-        return redirect('/todos');
+        try{
+            $todo->save();
+        }catch(\Exception $exception){
+            return redirect()->route('todos.index')->with('error', 'Đăng kí task thất bại');
+        }
+        return redirect()->route('todos.index')->with('success', 'Thêm task thành công');
     
     }
 
@@ -77,7 +83,11 @@ class TodosController extends Controller
     public function edit($id)
     {
         $todo = Todo::find($id);
-        return view('todos.edit', compact('todo'));
+        if($todo){
+            return view('todos.edit', compact('todo'));
+        }else{
+            return redirect()->route('todos.index')->with('error', 'Không tìm thấy task');
+        }
     }
 
     /**
@@ -90,11 +100,20 @@ class TodosController extends Controller
     public function update(TodoRequest $request, $id)
     {
         $data = $request->all();
-        $todo = Todo::find($id);
-        $todo->name = $data['name'];
-        $todo->description = $data['description'];
-        $todo->save();
-        return redirect('/todos');
+        try{
+            $todo = Todo::find($id);
+            if($todo){
+              $todo->name = $data['name'];
+              $todo->description = $data['description'];
+              $todo->save();  
+            }else{
+                return redirect()->route('todos.index')->with('error', 'Không tìm thấy task');
+            }
+            
+        }catch(\Exception $exception){
+            return redirect()->route('todos.index')->with('error', 'Update task thất bại');
+        }
+        return redirect()->route('todos.index')->with('success', 'Update task thành công');
     }
 
     /**
@@ -105,8 +124,19 @@ class TodosController extends Controller
      */
     public function destroy($id)
     {
-        $todo = Todo::find($id);
-        $todo->delete();
-        return redirect(route('todos.index'));
+        try{
+            $todo = Todo::find($id);
+            if($todo){
+              $todo->delete();  
+            }else{
+                return redirect()->route('todos.index')->with('error', 'Không tìm thấy task');
+            }
+            
+
+        }catch(\Exception $exception){
+            return redirect()->route('todos.index')->with('error', 'Xóa task thất bại');
+
+        }
+        return redirect()->route('todos.index')->with('success', 'Delete task thành công');
     }
 }
